+++ src/LowerTriangularArrays/LowerTriangularArrays.jl
@@ -4,9 +4,10 @@ module LowerTriangularArrays
 using DocStringExtensions
 
 # GPU
-import Adapt
+import Adapt: Adapt, adapt
 import GPUArrays
 import KernelAbstractions
+import ..Architectures: AbstractArchitecture, on_architecture, array_type, ismatching, CPU, GPU, architecture
 
 # NUMERICS
 import LinearAlgebra: tril!
@@ -14,7 +15,7 @@ import LinearAlgebra: tril!
 export AbstractSpectrum, Spectrum, resolution, truncation
  
 export LowerTriangularMatrix, LowerTriangularArray
-export eachharmonic, eachmatrix, eachorder
+export eachharmonic, eachmatrix, eachorder, orders
 export OneBased, ZeroBased
 export lta_view