+++ src/RingGrids/RingGrids.jl
@@ -10,9 +10,10 @@ import FastGaussQuadrature
 import LinearAlgebra
 export rotate, rotate!
 
-import Adapt
+import Adapt: Adapt, adapt, adapt_structure
 import GPUArrays
 import KernelAbstractions
+import ..Architectures: AbstractArchitecture, on_architecture, array_type, ismatching, CPU, GPU, architecture
 
 # ABSTRACT GRIDS
 export  AbstractGrid,
@@ -107,9 +108,9 @@ export zonal_mean
 # CONSTANTS
 const DEFAULT_NF = Float64
 const DEFAULT_ARRAYTYPE = Array
+const DEFAULT_ARCHITECTURE = CPU
 
 include("utility_functions.jl")
-include("architecture.jl")
 
 # GENERAL
 include("abstract_types.jl")