+++ src/RingGrids/architecture.jl
@@ -1,4 +1,0 @@
-struct DummyArchitecture end
-const DEFAULT_ARCHITECTURE = DummyArchitecture
-array_type(::Type{DummyArchitecture}) = Array
-array_type(arch::DummyArchitecture) = array_type(typeof(arch))