+++ src/RingGrids/grids/full_clenshaw.jl
@@ -46,3 +46,5 @@ get_lond(::Type{<:FullClenshawGrid}, nlat_half::Integer) = get_lond(FullGaussian
 
 # QUADRATURE
 get_quadrature_weights(::Type{<:FullClenshawGrid}, nlat_half::Integer) = clenshaw_curtis_weights(nlat_half)
+
+Adapt.@adapt_structure FullClenshawGrid