+++ src/RingGrids/grids/full_octahealpix.jl
@@ -38,3 +38,5 @@ get_lond(::Type{<:FullOctaHEALPixGrid}, nlat_half::Integer) = get_lond(FullGauss
 # QUADRATURE (use weights from reduced grids though!)
 get_quadrature_weights(::Type{<:FullOctaHEALPixGrid}, nlat_half::Integer) =
     equal_area_weights(OctaHEALPixGrid, nlat_half)
+
+Adapt.@adapt_structure FullOctaHEALPixGrid