+++ src/RingGrids/grids/healpix.jl
@@ -156,3 +156,5 @@ function each_index_in_ring!(   rings::AbstractVector,
         rings[j] = index_1st:index_end              # turn into UnitRange
     end
 end
+
+Adapt.@adapt_structure HEALPixGrid