+++ test/GPU/runtests.jl
@@ -2,6 +2,10 @@ using CUDA
 using SpeedyWeather
 using Adapt
 using Test
+using KernelAbstractions
 
 # transforms
 include("spectral_transform.jl")
+
+# kernels
+include("kernels_GPU.jl")