+++ test/differentiability/barotropic.jl
@@ -61,19 +61,15 @@
 
     autodiff(Reverse, SpeedyWeather.dynamics_tendencies!, Const, Duplicated(diagn, ddiag), Duplicated(progn, dprogn), Const(lf2), Duplicated(model, make_zero(model)))
 
-    # in the default configuration without forcing or drag, the barotropic model's dynamics_tendencies don't dependent on the previous prognostic state, but only on the diagnostic variables 
-    @test sum(to_vec(dprogn)[1]) ≈ 0 
-
-    # FD comparison not necessary, we have the exact values 
-    #function dynamics_tendencies(diagn, progn, lf, model)
-    #    diagn_new = deepcopy(diagn)
-    #    SpeedyWeather.dynamics_tendencies!(diagn_new, deepcopy(progn), lf, deepcopy(model))
-    #    return diagn_new
-    #end 
-    #
-    #fd_vjp = FiniteDifferences.j′vp(central_fdm(9,1), x -> dynamics_tendencies(diagn_copy, x, lf2, deepcopy(model)), ddiag_copy, progn_copy)
-    #
-    #@test all(isapprox.(to_vec(fd_vjp[1])[1], to_vec(dprogn)[1],rtol=1e-4,atol=1e-1))
+    function dynamics_tendencies(diagn, progn, lf, model)
+        diagn_new = deepcopy(diagn)
+        SpeedyWeather.dynamics_tendencies!(diagn_new, deepcopy(progn), lf, deepcopy(model))
+        return diagn_new
+    end 
+    
+    fd_vjp = FiniteDifferences.j′vp(central_fdm(9,1), x -> dynamics_tendencies(diagn_copy, x, lf2, deepcopy(model)), ddiag_copy, progn_copy)
+    
+    @test all(isapprox.(to_vec(fd_vjp[1])[1], to_vec(dprogn)[1],rtol=1e-4,atol=1e-1))
  
     #
     # horizontal_diffusion!
@@ -103,7 +99,7 @@
     # should be row-wise `model.horizontal_diffusion.impl .* model.horizontal_diffusion.expl`
     # for all variables that are diffused 
     diff_coefficient = model.horizontal_diffusion.impl .* model.horizontal_diffusion.expl
-    l_indices = [(1:l) for l=1:progn.vor[1].n]
+    l_indices = [(1:l) for l=1:progn.vor[1].spectrum.mmax]
     for (i,il) in enumerate(l_indices)
         @test all(real.(Matrix(dprogn.vor[lf1][:,1])[i, il]) .≈ diff_coefficient[i])
     end 
@@ -189,7 +185,7 @@
         return diag_copy
     end 
 
-    fd_vjp = FiniteDifferences.j′vp(central_fdm(5,1), x -> transform_diagn(diag_copy, x, lf2, model), ddiag_copy, progn_copy)
+    fd_vjp = FiniteDifferences.j′vp(central_fdm(12,1), x -> transform_diagn(diag_copy, x, lf2, model), ddiag_copy, progn_copy)
     
     @test all(isapprox.(to_vec(fd_vjp[1])[1], to_vec(dprogn)[1],rtol=1e-2,atol=1e-2))
 end